車機資料類型調整

+++ WebAPI/Models/Param/Input/IAPI_CarData.cs
@@ -35,7 +35,7 @@ namespace WebAPI.Models.Param.Input
         public double deviceVolt { get; set; }
         public double deviceLatitude { get; set; }
         public double deviceLongitude { get; set; }
-        public int deviceMillage { get; set; }
+        public double deviceMillage { get; set; }
         public int extDeviceStatus1 { get; set; }
         public int extDeviceStatus2 { get; set; }
         public string extDeviceData2 { get; set; }
